Remove all unused imports
