+++ ml_enabler/models/task.py
@@ -2,7 +2,6 @@ from ml_enabler.models.utils import timestamp
 from ml_enabler import db
 from ml_enabler.models.dtos.dtos import TaskDTO
 from sqlalchemy.sql.expression import any_
-import sqlalchemy
 
 class Task(db.Model):
     __tablename__ = 'tasks'
