+++ ml_enabler/services/imagery_service.py
@@ -1,6 +1,5 @@
 from ml_enabler.models.imagery import Imagery
 from ml_enabler.models.utils import ImageryNotFound
-from ml_enabler import db
 
 class ImageryService():
     @staticmethod
