+++ ml_enabler/services/project_service.py
@@ -2,8 +2,6 @@ from flask import current_app
 from ml_enabler.models.ml_model import Project, ProjectAccess
 from ml_enabler.models.dtos.dtos import ProjectDTO
 from ml_enabler.models.utils import NotFound
-from ml_enabler.utils import version_to_array
-from sqlalchemy.orm.exc import NoResultFound
 
 
 class ProjectServiceError(Exception):
