+++ ml_enabler/services/user_service.py
@@ -1,7 +1,4 @@
 from ml_enabler.models.ml_model import User
-from ml_enabler import db
-from ml_enabler.models.utils import NotFound
-import boto3
 
 class UserService():
