+++ requirements.txt
@@ -1,3 +1,4 @@
+pylint==2.6.0
 semver==2.10.2
 boto3==1.14.39
 pyproj==2.6.0
